Creados perfís para Amazon e local. Configurado o logging. Devolución dos POIs a partir do identificador do edificio externo (Situm).

+++ src/main/java/gal/caronte/sw/modelo/edificio/EdificioDao.java
@@ -6,6 +6,6 @@ public interface EdificioDao {
 
 	List<Edificio> getTodos();
 	
-	Edificio getEdificio(Integer idEdificio);
+	Edificio getEdificio(Short idEdificio);
 	
 }
